rerender answer list when new answer is added
