+++ client/src/components/QA/AnswersList.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 import axios from 'axios';
 import Answer from './Answer';
 import * as styles from './qanda.module.css';
@@ -6,25 +6,50 @@ import * as styles from './qanda.module.css';
 function AnswersList({ questionId, newAnswer }) {
   const [allAnswers, setAllAnswers] = useState([]);
   const [displayedAnswers, setDisplayedAnswers] = useState([]);
+  const [count, setCount] = useState(2);
   useEffect(() => {
     axios.get(`/qa/questions/${questionId}/answers`, { params: { count: 999 } })
       .then((res) => {
         setAllAnswers(res.data.results);
-        setDisplayedAnswers(res.data.results.slice(0, 2));
+        setDisplayedAnswers(res.data.results.slice(0, count));
       })
       .catch((err) => {
         throw new Error(err);
       });
   }, [newAnswer]);
 
+  useEffect(() => {
+    const a = allAnswers.slice(0, count);
+    setDisplayedAnswers([...a]);
+  }, [count, allAnswers]);
+
   function loadAllAnswers() {
-    // show all remaining answers
-    // confined to half the screen and scrollable
-    setDisplayedAnswers(allAnswers.slice());
+    setCount(allAnswers.length);
   }
   function collapseAnswers() {
-    setDisplayedAnswers(allAnswers.slice(0, 2));
+    setCount(2);
   }
+  const didMount = useRef(false);
+  useEffect(() => {
+    if (didMount.current) {
+      // console.log('gonna hit api with ans', newAnswer);
+      axios.post(`/qa/questions/${questionId}/answers`, JSON.stringify(newAnswer), { headers: { 'Content-Type': 'application/json' } })
+        .then(() => {
+          axios.get(`/qa/questions/${questionId}/answers`, { params: { count: 999 } })
+            .then((res) => {
+              setAllAnswers(res.data.results);
+              if (count !== 2) {
+                setCount(res.data.results.length);
+              }
+            })
+            .catch((err) => { throw new Error(err); });
+        })
+        .catch((err) => { throw new Error(err); });
+    } else {
+      didMount.current = true;
+    }
+  }, [newAnswer]);
+
   return (
     <div className={[styles.answersList, styles.scrollable].join(' ')}>
       <div>
@@ -32,8 +57,8 @@ function AnswersList({ questionId, newAnswer }) {
       </div>
       <div>
         {displayedAnswers.map((answer) => <Answer key={answer.answer_id} answer={answer} />)}
-        { (displayedAnswers.length === allAnswers.length && displayedAnswers.length > 2) ? (<input type='button' value='Collapse' onClick={collapseAnswers} />) : (<div />)}
-        {(allAnswers.length > 2 && displayedAnswers.length < allAnswers.length) ? (<input type='button' value='Load More Answers' onClick={loadAllAnswers} />) : (<div />)}
+        { (count > 2) ? (<input type='button' value='Collapse' onClick={collapseAnswers} />) : (<div />)}
+        {(allAnswers.length > 2 && count === 2) ? (<input type='button' value='Load More Answers' onClick={loadAllAnswers} />) : (<div />)}
       </div>
     </div>
